Extract podcast creation from dictionary to a method
So it can be used in unit tests
